fix:getUser 파일 별도로 만들어서 사용
